Fix showing the project overview if head is undefined, assume `master` then
If `head` is undefined, the ajax request listing the branches available
for PRs will fail because the `branch_to` argument will be `None` and
thus ignored while it's actually required. So instead of bailing out
just assume the head is `master`, the rest of the code should know how
to deal with this.

Fixes https://pagure.io/pagure/issue/3305

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>