Fix messed up custom fields in settings
The new tabbed layout of the settings page changed the way
we add rows to some of the settings that use rows, like priorities
and the custom fields. This new way used a hidden blank-field that
allowed us to duplicate. However, this field was sticking around when
submitting the form, and was causing the custom fields to not be
submitted propertly to the post call. This PR uses javascript to remove
the hidden field from the form before it is submitted.

Relates to #3366

Signed-off-by: Ryan Lerch <rlerch@redhat.com>