Tweak notification styles
Adds all notifications for a single request into a single
box that can be more easily dismissed when there are multiple
notifications. Also absolutely positions the notification so we
don't have to worry about if a page uses a sub-header or not (like
on the repo pages)

The notification body is now white, and the status of each notification
is signifed by the text colour and a small icon.

The welcome message still uses the alert-info colours.

Relates to #3347

Signed-off-by: Ryan Lerch <rlerch@redhat.com>