Support calls from jenkins indicating the build is started
Basically with this commit, if we receive a call from jenkins
about a certain job, we'll wait 5 seconds to see if this job
completes, if it does not we will flag the pull-request with
a flag saying that this build is pending (ie: in progress).
Then when the build finishes, we'll loop over all the tags of
the pull-request and see if one is pending and refers to the
same job, if it does we will just update this flag instead of
adding a new one.

Fixes https://pagure.io/pagure/issue/2833

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>