Refactor the pull_request.html template
Previously, the pull_request.html template was used for 3 different
pages: the PR page, the create PR page(displayed as a diff of two
branches when not logged in), and diffing two commits.

This PR pulls the actual rendering of the diff out into a new macro
file, that is shared between three new templates:

* repo_pull_request.html
* repo_compare_commits.html
* repo_new_pull_request.html

This does not change any of the style of the pages, it is simply
an abstraction of the templates to make the upcoming restyling
a lot easier.

Signed-off-by: Ryan Lerch <rlerch@redhat.com>