Re-generate the sha for every file.
Basically, we were instantiating the sha outside of the loop going
through the files. This meant that if we generated the sha for two
files, the sha of the second included the first one in it...

This commit fixes this be making sure we compute a new sha specific
for each file.

Fixes https://pagure.io/pagure/issue/3448

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>