remove nav-tabs border rules from pagure.css
Previously, we were using a rule in Pagure.css
to remove the bottom border of nav-tabs elements.
These nav-tabs were used in the repo-header tabs,
the browsing tabs, and the preview tabs for comments.

This commit removes this rule from the pagure.css, and
replaces it in the places we need it in the html with the
bootstrap utility class 'border-bottom-0'