remove css classes that hide the metadata forms in issues
the issue-custom form is no longer used, custom items are now
in the main metadata form. Also removed some referecnes to these
classes in the javascript.

the .issue-metadata-form is now hidden in the template by default,
using the bootstrap .hidden class. In fixing this, I found an issue
with the render_bootstrap_field macro, where it would always use the
issue-metadata-form class if the formclass parameter was set.