remove red-icon class
remove the red-icon class from the pagure.css. This was only being
used in the message about repo permissions being updated, and wasnt
even showing anymore because it was using the old openiconic icons.
Didnt replace it with the fontawesome lock, because we use this icon
elsewhere to signify a private repo or issue, and no real need for a
icon here tbh.