hack the code table to display diff highlights in PR compare
Previously the highlighting of diffs in the files changed
tab of the PR page used pygments. with pygments removed,
and using highlight.js everywhere else, diffs for commits
were no longer highlighted.

Using highlight.js in this view is a bit tricky without some
major work to changing how inline comments work #3488, so this
PR is a stop-gap to get highlighting of diffs here working until
we can fix this properly.

Fixes #3531