Move project options to their own template and start documenting them
This commit moves the entire section about project options into their
own template, which reduces the amount of code in the settings template.
In addition it adds a little bit of documentation about the different
options present and what they mean in a way that the user can show/hide
on demand.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>