reorganize the use of bootstrap files in themes
this adds the stock bootstrap to the base static file, and
imports it from there in the default theme. This also removes
the fedora-bootstrap js files which were not actually providing
anything more than the stock bootstrap js, and makes the two
fedora themes use the js from the base static directory.

fixes #3575