remove jquery-ui
This removes jquery-ui from pagure. The only place where we were
still using this was the api docs page for the accordions. After
removing this, the api docs still render ok, just with no collapse.

This will need to be re-thought to make this page a lot better, but
keeping this task for post 5.0

Fixes #3583