Integrate repoSpanner
This allows configuring new repositories to be created on repoSpanner, and allows
existing ones to be migrated over.
Support is included for multiple regions/setups, although migrating between regions
is not supported (technically nothing limiting, but not required for first version).

The actual repoSpanner parts require a repoSpanner enlightened libgit2 build.

The test suite requires the patched libgit2 and a repoSpanner daemon binary.
This binary can either be in /usr/bin/repospanner, or in the top level directory.
If this binary can't be found, the test will be marked as Skipped.
In addition, it requires an /etc/hosts entry with:
127.0.0.1	nodea.regiona.repospanner.local

Signed-off-by: Patrick Uiterwijk <patrick@puiterwijk.org>