fix styling of input-group-addons
WE previously used bootstrap input-group-addons to
style up a few elements, specifically the git and ssh urls
in the clone dropdown, and the listing of api keys in the
settings for a project and a user.

Since we were using pre-release versions of bootstrap 4 earlier,
bootstrap changed slightly how these are marked up in newer versions.

This commit changes them so they now are visually the same as
with older versions.

Fixes #3616