introduce the browseheader_message macro
Introduces a new macro for themes, the browseheader_message.
This allows a theme to define the message that is shown on the
browse pages. In the default theme, this message was and is only
shown on the browse projects.

This also updates the srcfpo theme to use this new macro, and remove
the overridden template that was being used to do this.

Fixes #3617