Move to BaseRunner hooks architecture
This starts moving repository hooks over from a set of symlinks into
the git repos' hooks/ folder into the files over to a Runner class
per plugin.

This requires less constant meddling with files, and allows a single
hook to perform multiple tasks at the same time.

To do after this is changing lib.git.TemporaryClone.runhook to call
run_project_hooks and removing lib.repo.PagureRepo.run_hook.
Also to do is migrating the rest of the in tree hooks over to the
Runner architecture.

Signed-off-by: Patrick Uiterwijk <patrick@puiterwijk.org>