Add hooks to repoSpanner setups
This adds a hook runner script, which needs to be uploaded by the admin
with the admin CLI upload-repospanner-hooks script.
This returns a hook ID that is inserted into the configuration, after
which it is automatically added to new repositories on repoSpanner.

To make sure that the hooks are created for all repos, an
admin ensure-project-hooks cli task is also added to make sure that all
projects (Also those not on repoSpanner) have their hooks installed.

Signed-off-by: Patrick Uiterwijk <patrick@puiterwijk.org>