Make the term we call a Project configurable by the theme
By default, Pagure refers to a code project or repository throughout
as a "Project". In some cases, when running an instance of Pagure,
it would be desirable to refer to this as something else, like simply
a "repository" or in the case of src.fedoraproject.org, a "package".

This change abstracts out this term, (and the icon used to represent
a 'project') and adds the ability to set two new optional variables
in the theme.html to change these to whatever the theme maker wants:

* 'projectstring' -- default:'Project'
* 'projecticon' -- default:'fa-calendar-o fa-rotate-270'

This also updates the theme for srcfpo to change these to 'Package' and
'fa-archive'