Implement initial keyhelper
This can be configured in sshd as the AuthorizedKeysCommand to dynamically
get the keys from the Pagure database on the fly.

The current version is very naive and just loops over all the users if
user lookup isn't used.
This should be improved in the future by making the user SSH keys be stored
in the schema like DeployKey's, where each key has its own entry in a table
and we can look them up by key ID.

Signed-off-by: Patrick Uiterwijk <patrick@puiterwijk.org>