Support remote PR and running pagure-ci
Remote PR in pagure have no project_from since they are remote but up
until now pagure-ci's code was relying on the fullname of the
project_from. With this commit we add support for remote PR which, when
trying to trigger a new CI run, can specify a PR UID instead of a project
fullname. The logic will then retrieve the corresponding PR, check if it
is a remote PR or not and depend on this use either the project_from or
the project_to field of the pull-request object.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>