Fix syntax highlighting of files.
This updates the file display to use just highlight.js
Previously, highlight.js was run over the old style
codetable that was generated for when we used pygments,
and this was causing each line to be rendered seperately.
This caused hilight.js to determine some strange filetypes
for these fragments, and typically hilight them as bash output.