Rework how we display loading of the new PR dropdown
This reworks the display of the new-pr dropdown,
so we show loading animations for both the async
requests, and display the results in a consistent manner.

Previously, whatever request finished first, those results
would be at the top.

This does not change the ordering of the results, just how
the dropdown is displayed.

Fixes: #3812