Make the mirroring feature work with older git
The GIT_SSH_COMMAND environment variable was added in 2.10 but we
want pagure to work with older git version, so we are switching
to GIT_SSH with a helper script generated on the fly.

In addition, this commit adds the logic to pull all the branches stored
in pagure so when we push --mirror they are all synced.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>