Make sure that if a user is direct committer, they can always commit
This fixes the case where the user is in an EXTERNAL_COMMITTER group that is
excluded from a specific repository, but they are actually a direct committer
(or in a group that is a direct committer) for that repository.

Example for Fedora Dist-Git: commit for "rpms/firefox" if one of the firefox
maintainers were provenpackager.

Signed-off-by: Patrick Uiterwijk <patrick@puiterwijk.org>