Don't use unicode literals for git_multimail
It uses a literal as a type() argument, which doesn't work with unicode.
This is also a file we forked from upstream, so we shouldn't be changing
it up.
This could perhaps break python3 compatibility, but we should just either
work with upstream to fix that, or rebase on the latest upstream if they
already have.

Fixes: #3884
Signed-off-by: Patrick Uiterwijk <patrick@puiterwijk.org>