Fix the PYTHONPATH for the dev_worker_1 container
When trying to create new project using API, with the `create_readme=True`
parameter, there was an ImportError: No module named `pagure.lib`, in
the pre-receive hook of the new project.

In dev_worker_1 container the PYTHONPATH was pointing to `.`.
The hooks were probably not being run from the workdir set in the
Dockerfile. Now it points to `/code` and the error is gone.

Merges #3944