Drop specifying redis all the time, but allow not calling webhooks
The redis argument that was provided was not used in its original purpose
which was to only send notifications to webhook if redis was properly
configured, but it's now all handled by the workers and celery.
However, in one place, the default hook, it was use to allow by-passing
sending webhook notifications.
So to keep this behavior in place, we introduced a dedicated argument.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>