Fix notifications and refreshing the cached merge status upon updates
Basically, before this commit, when we pushed a new commit to the git
repo, we would update all the open pull-requests involved with this
branch so that we can link them to tickets, then we would refresh all
the open pull-requests of the project (ie: clear out their cached merge
status).
With this commit, before we link PRs to ticket, we refresh them, so it
sends notifications about rebase to the configured bus (if any) and
comments about it in the UI. This process also refreshes the cached
merge status.
So we have to be careful not to clear this updated cached merge status
later otherwise we updated it for nothing.
So this commit ensures we only clear cached merge status of PRs we have
not refreshed manually.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>