Include whether the PR passed the threshold or not in the API data
Basically, if the project requires a minimal score to merge PR,
the data returned by the API will be either a ``True`` or
``False``.
If the project does not enforce any score to merge the PR, the
data returned by the API will be ``None``.

Fixes https://pagure.io/pagure/issue/4014

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>