Specify the home when doing pushes
This is needed on deployment where all users are not using the same user
on the system, for example on src.fedoraproject.org.
In this case, the hook is ran as a dedicated user which is likely
different from the gitolite user.
So specify that the HOME folder is the gitolite's so it gitolite can
find where to put the logs.