Prevent double click from showing two input form
Previously, if the server was busy and you would click twice on the
button to add an inline comment, you could end up with two (or more!)
input forms showing up for this line in the diff.

This effectively disables the button until it is known that the user can
now asks for more forms to submit more comments.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>