Change the way votes are recorded
Before this commit there would only be one vote per person and we would
sum up the values for each, so if you voted once -1 and once +1 your
action would result in a 0.
With this change, we record your last vote, so if you voted once -1 and
then +1, your vote would lead to a +1.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>