PV endpoint: check score after the merge_status eval
This change makes sure a PR got a merge status other than
'unknown' when minimal score to merge is set in the settings.
Indeed, if set, the score is checked first, and merge_status is
not set if score no reached. Even if the score is not reached
it makes sense to known the mergeability.