Change "created by" to "maintained by" in repo info
This changes the string in the repo info page that
displays the main admin of a repo from 'created by' to
'maintained by'. This change is due to the fact that ownership
of a repo can be changed, and the current main admin my not
be the user that actually created the project.

Also adds a second line that shows the creation date of a
repo. Also changes the owner string in userdash projects
to maintainer, and the main admin in the repo acls list
on repoinfo to maintainer.

Fixes #4055