Allow filtering user's PR by time information
This commit adds support for filtering the PR filed or actionable for an
user using time information such as the created date, the updated date
or the closed date.
The format supported it since..until. If the string provided does not
include a '..', it is assume that only the since value was specified.

Fixes https://pagure.io/pagure/issue/4065

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>