Cascade deleting flags when tokens are deleted
A project is deleted, if may have PRs with flags, these flags
have been added via the API via an API token (potentially linked
to the project).
When the project is deleted the API token gets deleted and the
corresponding flags should be deleted as well, which this commit
allows.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>