Add project connector api endpoint
This endpoint aims to provide a get access to connector
tokens that are:
- project hook token
- project's user (caller) API token

I call them connector data as those tokens are the
mandatory data an application should have to verify project
events received via the Web hooks but also to perform actions
via the API to a given project.

The endpoint is only available to project's owner and admin
users with a 'modify_project' ACL token.

Thanks to this endpoint, a user (or an application) added as
admin on one or more projects, and via its user token, would
be able to read connector data on those projects and then
received events and perform API call on those same projects.