Fix not showing the edit and delete buttons when they won't work
If there is no identifier (which is the case when the comment is added
using ajax rather than being received by SSE), the edit and delete
buttons on the comment won't work, so there is no point in showing them,
the users will just have to refresh the page to see them.

In addition, this commit removes the link to the comment itself, for the
same reason.

Fixes https://pagure.io/pagure/issue/4210

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>