Write the ssh key to disk rather than relying on /dev/stdin
This is required to get the tests to pass in a container where /dev/stdin
is not available inside the process running the tests.
So rather than relying on /dev/stdin, we just write the key to disk in a
temp folder that is cleaned up automatically once we're done checking if
the ssh key is valid or not.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>