Fix runtests.py so it works in containers
Basically in a container we want to mount the results directory
so the test results are exported outside of the container.
This means, the results directory cannot be deleted or created
from inside the container. So if we can delete it, let's just
remove all the files it contains so we start fresh.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>