CI: fix git-log to not page output
Sometimes in strange tty situations (in a container that was started by
an ssh session, within a jenkins workspace for example). Git gets
confused about what it should do with its output and tries to be helpful
about piping it to a pager. It shouldn't do that in the CI context
because that blocks execution.

Signed-off-by: Brian Stinson <brian@bstinson.com>