Load all the tickets for the depending/blocking fields
The ajax call had not been adjusted for the change in the pagination
of the endpoints so it was returning only some of the tickets instead
of all of them.
This commit adjusts this. We may need to optimize this for large projects
but this can be done in another change.

Fixes https://pagure.io/pagure/issue/3939

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>