Increase logging and support the reference returning a commit
We had a case where the reference given wasn't pointing to a tag
but to a commit. Which as a matter of fact, is fine since when
we are given a tag we want to go to the underlying commit.
So if we're given a commit, keep going.

Fixes https://pagure.io/pagure/issue/4278

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>