Fix adding multiple tags to an issue
In commit 4d961f80509c9279d563674132dd51dbb040f3e3 we forbid the use of
comas in tag names. This is fine, except when seeing and updating an
issue, tags are sent as a coma-separated list. That commit thus broke
adding multiple tags to a single issue, kinda annoying.
To fix this, this commit introduce a new regex which is basically the
same regex as before but which allows comas in there. Thus allowing to
have multiple tags in an issue without allowing comas in the tag names.

To ensure we don't break this by accident in the future again, we're
also adding a couple of unit-tests checking that the behavior is
consistent with the expectations.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>