Add support for !owner to the API listing projects
When querying the list of projects one can filter the projects returned
by their main maintainer, using the filter argument `owner.
With this commit, if the filter argument starts with the `!`, instead
of filtering for project having this person as maintainer, it will allow
to filter for project not having this person as main maintainer.

Fixes https://pagure.io/pagure/issue/4312

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>