Allow linking issues to PRs in the initial comment of a PR
Currently, we allow linking issues to PRs in commit messages but there
is no reason we cannot link issues to PR via the initial comment of a
PR.
This commit adds support for this. One can now link an issue with a PR
in the initial comment, when opening the PR as well as when editing the
initial comment.

Fixes https://pagure.io/pagure/issue/4309

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>