When calling flask.abort() specify the description
In the most recent release of werkzeug (0.15.0) some changes around the
handling of 401 exceptions has changed the order of the arguments.
Thus making 'flask.abort(401, "something")' not considering "something"
to be the description of the exception.
With this commit we specify in all calls to flask.abort() what is the
description of the error.
This way, there is no longer any confusion for flask or werkzeug as to
what we want the description to be.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>