Do not link on the "Star" button if the user is not authenticated
Otherwise, people who are not authenticated are able to click on the
button. This will redirect the user to the login page, then back into
the star project endpoint, but this endpoint only accepts POST requests
while redirect can only be done via GET requests.
So the user ends up with a page showing a 405 (method not allowed) error.
By dropping the link for not-authenticated users, we solve this issue.

Fixes https://pagure.io/pagure/issue/4372

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>