Fix the new PR drop-down button
Currently we only populate the new PR drop-down button if the user
is a maintainer, however, if the user has a fork, the drop-down
button should work.
So instead of preventing the entire button from working, we've
adjusted so that only the main project is considered if the user
is a maintainer and the fork is considered everytime otherwise.

Fixes https://pagure.io/pagure/issue/4310

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>