Fix deleting a project if creation failed
Basically, if a project fails to be created we want to have it
removed from the DB. So far, we only reverted and removed the
project from the DB if the repositories were found on disk. With
this commit, we'll remove the project from the DB for any
exception raised.

Fixes https://pagure.io/pagure/issue/4384

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>