Couple of fixes for the mirroring-in feature
This commit does a couple of fixes in the code used for the mirroring-in
feature.
It ensures any exception thrown by pagure.lib.git.mirror_pull_project are
caught and properly logged. Otherwise, the script dies in the middle of
the processing.
It also checks that there is a remote before doing anything and if there
is no remote, it just bails.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>