Fix linking to issues or PRs when pre-viewing a comment
Basically the old mechanism was retrieving an user '' which was not
considered to be None, so when searching for the matched issue/PR,
none were found and thus the text version was returned.
With this change, we ask that username, project name or project
namespace are all either something or None. The code is then
able to find the issue/PR referenced in a comment when previewing
it.

Fixes https://pagure.io/pagure/issue/4319

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>