Call the source package info in mdapi
We were calling mdpi for the package info, but the dist-git repo name
corresponds to the source package name, not the actual package name.
So, we would not find anything for packages such as python-arrow
whose source package name is python-arrow but the packages are
python2-arrow and python3-arrow.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>