Add support for allowing the maintainers of the target project rebase
If the person opening the PR checks this checkbox, pagure will by-pass
the auth check when pushing to the project the PR originates from and
thus allow rebasing.
Otherwise, rebase would only be allowed on project the maintainer has
commit access to, which is likely not the case for most forks.
In addition, since we do not know what the fork is used for, it could be
that the fork is used by the contributor as a basis for their own
deployment, we want the permission to rebase the fork to be explicit to
avoid changing the fork without the contributor's knowledge.

Fixes https://pagure.io/pagure/issue/4456

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>