Change the way we sanitize the blamed view
We used to sanitize the entire content of the filter blame_loc which
builds the html table with the git blame view.
This sanitization removes all src in img tag that do not point to this
pagure instance, in other words, it removes all the avatars making the
page not really nice to see.

In this commit, we clean each line in the file that is being blamed
while building the html table instead of cleaning the entire table at
once.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>