Sync improvements to syntax highlighting from highlightjs-rpm-specfile
This file was upstreamed in a different form as highlightjs-rpm-specfile,
and that version was improved to cover many of the new features in
RPM 4.14 and RPM 4.15.

This change syncs those improvements back into our version.

In addition, the default syntax highlight override now uses the upstream
name for the rpm specfile syntax highlighter.

Signed-off-by: Neal Gompa <ngompa13@gmail.com>

Merges #4570

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>