Fix the logic around interacting with read-only databases in hooks
GIT_HOOK_DB_RO is set to True when the git hooks have a read-only
access to the database. When they have a read-only access to the
database, they can't edit it and thus need to send instructions
interacting with it to a worker who can do the edit asynchronously.

Turns out we had the logic exactly opposite as to what it should be.
This commit fixes that.
Now how did we get away with this for so long is a mystery.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>