swap "Add" and "Cancel" button ordering for access management
Prior to this change, the "Cancel" button was much too close to the
place where users click to select an access level. This made it too easy
for users to accidentally click the "Cancel" button after quickly
clicking on all the other interface choices.

Move the "Add" button to the left so that it is directly below the
dropdown input field.

Merges https://pagure.io/pagure/pull-request/4665