Fix checking if the user is a committer of the repo the PR originates from
There are a few cases where if the PR originates from the same repo that it
targets the ``project_from`` attribute of the ``PullRequest`` object is
``None``.
Therefore, we need to check if the user is a committer of either the
``project_from`` if there is one and otherwise the target repo.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>