Only consider the most recently active branches for new PR
Instead of considering all the branches in the git repository when we
are looking for the branches available to open new PRs from, we are
now considering only the 8 latest active branches. This should
optimize the "New PR" drop-down button in the UI quite significantly in
large repositories.

Fixes https://pagure.io/pagure/issue/4587

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>