Add the GET project/webhook/token endpoint
This endpoint allows a project's collaborator to access the
project webhook token. It is useful in case a collaborator
user want to react to project events via the webhook system
because it will need to validate the webhook payloads.

In particular, third party applications that need
to react to multiple projects events wont need to be configured
with a long list of token (one for each project) to be able to
validate event's payloads. Instead the application will just
need to access the projects' endpoint to figure out webhook
tokens.