Fix the regenerate-repo actions
This action allows to re-generate the JSON representation of each issue
or PR in a repo in their corresponding git repo.
The call they made to pagure.lib.git.update_git contained an argument
"repofolder" that has been removed a while back and it looks like these
lines were not updated for this.

This commit fixes that

We're also changing the message displayed to the user so it indicates the
action is ongoing (since it's an async task) rather than it being done.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>