Order the user and groups names in the access lists
When re-generating all the JSON representation of the tickets or PR
in their corresponding git repo, if the order varies, we end up
re-generating JSON blob for a similar content.
By keep the order of the user and group names sorted, we ensure that
we will not re-generate JSON blob for no reasons (at least for these
fields).

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>