Use assertRegex instead of assertIn to avoid non-determinism failures
Depending on the system load, we might see messages containing "just
now" or "seconds ago".  For example:

  "Merged just now..."

vs.

  "Merged seconds ago..."

If we only look for "just now", we are obviously going to fail when
the system takes some time to process a request, and "seconds ago" is
displayed.  This commit addresses this problem by using a regex when
matching the text, and expecting both patterns.

Fixes #4751

Signed-off-by: Sergio Durigan Junior <sergiodj@sergiodj.net>