Fix taking into account the blacklisted patterns
Basically, we were checking if a new project matches the blacklist
before taking into account arguments such as: is user namespace
enforced, or is the project private and so on, so patterns in the
blacklist such as docs/* would not be matched if the user was named
'docs' and user_namespace was enabled on the instance.

We're adding tests to this to ensure we don't run into this again.

Fixes https://pagure.io/pagure/issue/4726

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>