Make the reset button not the default action of the form
Basically, if you enter a value in the form and press the "enter" key,
the first/default submit button of the form is triggered. This meant in
this case that the "reset" action was the one triggered by default.

With this change, we make the reset button be of type "button" instead
of being of type "submit", thus making it not the default action while
still looking pretty.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>