Omit breaking original comment format in reply
This fixes replying to a comment with block quotes and other formatted
text by prefixing each line in the original Markdown text with `>` to
create additional quotation level. Previously, this was done on the
rendered Markdown, but it removed the formatting.

In some cases, the Markdown rendering can be still broken, for example
code blocks using triple backticks. From Python-Markdown documentation
[1]:

> Fenced Code Blocks are only supported at the document root level.
> Therefore, they cannot be nested inside lists or blockquotes.

[1] https://python-markdown.github.io/extensions/fenced_code_blocks/

Fixes #2479

Signed-off-by: Lukas Holecek <hluk@email.cz>