Handle attempts to create an existing git tag
If someone attempts to create a git tag that already exists, pygit2
will raise an error that we are now catching.
To distinguish if the git tag that was asked to be created was created
or not, we also include in the JSON returned a ``tag_created``
field that is a boolean indicating if the tag was created or not
(not created in this case means that it already existed).

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>