Make sure orphaned & retired packages can be claimed
When the package is orphaned, the "take" button shows as disabled in the UI,
until we check if the package is orphaned and retired or just orphaned.
Except that in the case where the package is orphaned and retired,
we re-purpose that button to send the users to the releng issue tracker to ask
for the package to be un-retired/un-blocked,
but we did not re-enable the button.

This led to a button being re-purposed but not re-actived,
and thus clicking on the button did not have any effect.

This commit fixes that by making sure that we always re-enable the button,
either because the package is orphaned and not retired and packagers can just
adopt it, or because the package is orphaned and retired and by clicking on
the button packagers will be sent to the releng issue tracker where they'll be
able to ask for the package to be un-blocked.

Signed-off-by: Miro Hrončok <miro@hroncok.cz>

Co-Authored-By: Pierre-Yves Chibon <pingou@pingoured.fr>