Allow deploy keys to commit to the doc git repository of a project
While stored in a different git repo, docs are part of the sources of
a project and should not contain any privileged information unlike
the tickets or requests git repositories.
Thus we are fine with granting commit access to deploy keys to the
doc git repository.

Fixes https://pagure.io/pagure/issue/4763

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>